Subtraction game
Hi!

Don't feel the need to read this unless you're interested -- this will mainly just be helpful for me when looking back.

Attached are the html and js files for the subtraction game. I have the bare bones of the game down, and I am in the process of switching up the structure so that instead of producing random numbers, it automatically generates ten problems and cycles through them when the user hits "next problem" with the intention that after 10 problems (we can easily change that number), it can tell the user which specific problems they got correct and which ones they got incorrect. Right now, all of the code for that is in place, but I am trying to figure out what to link the "next problem" button to so that it keeps the page the same but changes the problem, and changes what the problem prints onto the screen. I think the issue that I keep running into is that html doesn't cycle back to update parts of the page as the value for the variables change (if that makes sense), rather it reads through the code linearly, so when I link the button to a method it just does whatever that method tells it to without the rest of the page.